requirements.txt
flask>=3.1.1
flask-sqlalchemy>=3.1.1
flask-dance>=7.1.0
flask-login>=0.6.3
flask-wtf>=1.2.1
flask-limiter>=3.5.0
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
oauthlib>=3.2.2
pyjwt>=2.10.1
pillow>=11.2.1
sqlalchemy>=2.0.41
werkzeug>=3.1.3
requests>=2.32.4
cryptography>=45.0.4
pypdf2>=3.0.1
email-validator>=2.2.0
redis>=5.0.0
bleach>=6.1.0
python-magic>=0.4.27